pyproject.toml
[project]
name = "wx_bot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "huiwan_code", email = "huiwancode@gmail.com"},
]
dependencies = [
    "wechatpy==1.8.14",
    "Flask==1.1.1",
    "werkzeug==2.0.3",
    "Flask-RESTful==0.3.8",
    "Flask-Redis==0.1.0",
    "gunicorn==20.0.4",
    "langchain==0.1.13",
    "jinja2==3.0.3",
    "itsdangerous==2.0.1",
    "cryptography>=42.0.7",
    "baidu-aip>=4.16.13",
    "chardet>=5.2.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"
